requirements.txt
# FastAPI & Web Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Supabase
supabase==2.3.0

# Embeddings & ML
sentence-transformers==2.3.1
torch==2.1.2
transformers==4.36.2

# Data Processing
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Requests
httpx==0.26.0
requests==2.31.0

# Utilities
tqdm==4.66.1
numpy==1.24.3

# Optional: Streamlit for demo
streamlit==1.30.0
openai>=1.0.0
